check whether entered num is perfect number or not
Ask the user to enter an integer number. Check whether the number is
positive or not. If the entered number is positive check whether the entered
number is perfect number or not otherwise display proper error message. (a
perfect number is a positive integer that is equal to the sum of its positive
divisors, excluding the number itself. )